feat: support calculations in amount input field
Using addition, subtraction, multiplication or division in input will return the result of the expression.

Input validation was also updated to support mutliple numbers / terms in the input field.

If expression is invalid (e.g. two numbers without an expression between) then the fallback is to return the first valid number or 0.